[flink] Fix checkstyle FileLength Violation
